tty: Bail out if the TTY is not accessible
One common problem when accessing the TTY is that you cannot access it
as regular user. The expected behavior is that the TTY terminal
application errors this out to the user. TIO is simply looping trying to
endlessly access the TTY, leaving the user in the dark about what's
going on.

The underlying issue is that the access checking is done in a void
returning function (tty_wait_for_device()) that is basically doing
nothing when the (access()) function is failing (because the TTY is not
accessible).

Move the check for the TTY accessibility at the start, even before
configuring the TTY because why do we need to configure a TTY that we
cannot even access?

Signed-off-by: Carlo Caione <ccaione@baylibre.com>